Move Dashboard model related tests to appropriate file
